Fix errors when accessing env version data (#227)
* Fix null reference issue in VerifyVersionsFile

Added a check for the existence of environment versions in
`allVersions.EnvVersionGrouped`. If the entry is missing,
`envVersions` is initialized to an empty array to prevent
null reference exceptions later in the code.

* Fix envVersions handling in EntryHelpers

Updated the logic to skip iterations when the envLower key is not found in the EnvVersionGrouped dictionary, removing incorrect initialization of envVersions to an empty array.