Add support for common artifacts in pipelines
This commit introduces functionality to handle common artifacts in the `BaseNukeBuildHelpers` class. It adds the ability to upload common artifacts with a custom filename option in the `AddReleaseAsset` and `AddReleaseCommonAsset` methods. The logic for creating and uploading these artifacts has been updated to use a common output directory instead of a temporary directory.

New methods `CreateUploadCommonArtifacts` have been added to both the `AzurePipeline` and `GithubPipeline` classes to facilitate the upload of common artifacts during various stages of the pipeline (pre-test, build, post-test, and publish). Additionally, artifact naming conventions have been updated to incorporate the `$common` identifier for better management of common assets throughout the build and release process.